requirements.txt
pyyaml
click
nbconvert
ipython
